Handle .init_array and .fini_array symbols.

+++ src/elf/string_table.rs
@@ -66,8 +66,8 @@ impl StringTable {
         self.names.get(&name).map(|(i, o)| Entry { index: *i, offset: *o, new: false })
     }
 
-    pub fn sh_name(&self, name: impl Into<Name>) -> usize {
-        self.names.get(&name.into()).unwrap().1
+    pub fn sh_name(&self, name: impl Into<Name>) -> u32 {
+        self.names.get(&name.into()).unwrap().1 as u32
     }
 
     pub fn name(&self, offset: usize) -> Option<&Name> {
